tests update

+++ tests/Functional/Repository/test_user_repository.py
@@ -1,6 +1,8 @@
 from django.contrib.auth.models import User
 from django.test import TestCase
 from dashboard_meet_je_stad.repository.user_repository import UserRepository
+import os
+from django.apps import apps
 
 
 class TestUserRepository(TestCase):
@@ -26,9 +28,11 @@ class TestUserRepository(TestCase):
         user = self.user_repository.create(user, 'secret')
         user_id = user.id
 
-        self.assertEqual(user.dashboarduser.get_sensor_id(), 840)
-        self.user_repository.save_dashboard_user(user, 1196)
-        user = self.user_repository.get(user_id)
-        self.assertEqual(user.dashboarduser.get_sensor_id(), 1196)
-        self.user_repository.delete(user)
-        self.assertEqual(user.id, None)
+        if os.path.isfile(os.path.dirname(apps.get_app_config('dashboard_meet_je_stad').path)
+                          + '/tests/data/Stations tbv Mailchimp.xlsx'):
+            self.assertEqual(user.dashboarduser.get_sensor_id(), 840)
+            self.user_repository.save_dashboard_user(user, 1196)
+            user = self.user_repository.get(user_id)
+            self.assertEqual(user.dashboarduser.get_sensor_id(), 1196)
+            self.user_repository.delete(user)
+            self.assertEqual(user.id, None)
